Create readme.md
Get your lasso out — the second step in the data science method is data wrangling! Data wrangling involves taking raw data and preparing it for processing and analysis. In this subunit, you'll get an introduction to this important topic, which will also be explored in its own unit later on in this course.